requirements.txt
beautifulsoup4==4.6.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
Flask==1.1.1
Flask-Cors==3.0.8
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
lyricsgenius==1.8.2
MarkupSafe==1.1.1
matplotlib==3.1.2
numpy==1.17.4
oauthlib==3.1.0
pandas==0.25.3
PyMySQL==0.9.3
pyparsing==2.4.5
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.11.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
scipy==1.3.3
seaborn==0.9.0
selenium==3.141.0
six==1.12.0
termplotlib==0.2.3
tweepy==3.8.0
urllib3==1.25.7
Werkzeug==0.16.0
